requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

google-adk==0.1.0

google-cloud-aiplatform>=1.87.0
google-cloud-firestore>=2.16.0
google-cloud-pubsub>=2.21.0

google-api-python-client>=2.157.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0